Getting JSON from SoundCloud now

+++ db/migrate/20110708084804_create_tracks.rb
@@ -2,10 +2,10 @@ class CreateTracks < ActiveRecord::Migration
   def self.up
     create_table :tracks do |t|
       t.string :soundcloud_url
-      t.decimal :position
-      t.decimal :beatport_genre_id
+      t.integer :position
+      t.integer :beatport_genre_id
       t.string :mix_name
-      t.decimal :beatport_track_id
+      t.integer :beatport_track_id
       t.string :name
 
       t.timestamps
