+++ db/migrate/20110708122821_create_artists.rb
@@ -2,7 +2,7 @@ class CreateArtists < ActiveRecord::Migration
   def self.up
     create_table :artists do |t|
       t.string :name
-      t.decimal :beatport_artist_id
+      t.integer :beatport_artist_id
 
       t.timestamps
     end
